Cleanup and fixes

+++ emulator/src/sample.rs
@@ -1,6 +1,5 @@
 // Type used to store one sample containing multiple channels in
-// Switch to u16 later for 16 channels
-pub type Sample = u8;
+pub type Sample = u16;
 
 // Buffer containing samples
 pub struct SampleBuffer<'a> {
